services/orchestrator-agent/requirements.txt
flask==3.0.0
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
